requirements.txt
supabase
pandas
anthropic
google-generativeai
groq
scikit-learn
matplotlib
seaborn
python-dotenv
tabulate
